Oauth and better exception handling
Added support of the Tictail OAuth, so it's now possible to use the API
to optain an AccessToken based on a AuthCode (this is used when building
external tictail apps).
Added a TictailException for better handling of errors from the Tictail
API, with better messages.
Added additional unittests, espcially for the new OAuth and the
TictailException.